SEQUOIADBMAINSTREAM-4648 传参类型由uint16改成int32
git-svn-id: http://192.168.20.11/sequoiadb/trunk@42363 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/sequoiafs/sequoiaFS.cpp
@@ -228,7 +228,7 @@ error:
 }
 
 INT32 getCollection(sdbCollectionSpace &cs, const CHAR *clName,
-                    sdbCollection &cl, UINT16 replsize)
+                    sdbCollection &cl, INT32 replsize)
 {
    INT32 rc = SDB_OK;
